+++ internal/pathutil/pathutil.go
